+++ CHANGELOG.md
@@ -3,6 +3,11 @@
 The `pycldf` package adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 
+## [Unreleased]
+
+Fixed issue where the wrong value was assigned to ORM objects' `description` attribute.
+
+
 ## [1.40.2] - 2024-12-23
 
 Fixed issue whereby valid datasets could not be loaded into SQLite because the `database` module