+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 The `pycldf` package adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 
-## [Unreleased]
+## [1.40.3] - 2025-01-03
 
 - Fixed issue where the wrong value was assigned to ORM objects' `description` attribute.
 - Better error message when validating source with trailing separator.